% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expit.R
\name{expit}
\alias{expit}
\title{The expit or logistic function}
\usage{
expit(x)
}
\arguments{
\item{x}{A numerical scalar or vector.}
}
\value{
The value or vector of expit(x).
}
\description{
The expit or logistic function
}
\details{
This function is written on the basis of the built-in hyperbolic tangent function \code{tanh}.
}
\examples{
x <- 1:3
expit(x)
}
\seealso{
\code{\link[base]{tanh}}
}
