% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IC.R
\name{IC}
\alias{IC}
\title{The objective function (an information criterion) for optimization in WOCR models.}
\usage{
IC(lambda, d, gamma, U, y, model = "RR.d.lambda", a = 20,
  criterion = NULL)
}
\arguments{
\item{lambda}{The tuning parameter(s) to be optimized. It is 1-D in models \code{RR.d.lambda}, \code{RR.gamma.lambda},
\code{PCR.d.c}, \code{PCR.gamma.c} and 2-D in models \code{PCR.d.a.c} and \code{PCR.gamma.a.c}.}

\item{d}{The vector of singular values of the design matrix \eqn{X}.}

\item{gamma}{The vector of coefficients, consisting of the inner product of each orthogonal component and the repsonse \eqn{y}.}

\item{U}{The \eqn{U} matrix whose columns are the orthogonal components.}

\item{y}{The response vector \eqn{y}.}

\item{model}{Specifies the WOCR model to be fitted. Six choices are possible: \code{RR.d.lambda},
\code{RR.gamma.lambda}, \code{PCR.d.c}, \code{PCR.gamma.c}, \code{PCR.d.a.c} and \code{PCR.gamma.a.c}.}

\item{a}{The fixed shape parameter \eqn{a} in models \code{PCR.d.c} and \code{PCR.gamma.c} only. By default, \eqn{a} is set to 20.}

\item{criterion}{Specifies the information criterion to be minimized. The appropriate choice is one from \code{"AIC"}, \code{"BIC"}, or \code{"GCV"}.
By default, \code{criterion} is left as \code{NULL} and the recommended one is used.}
}
\value{
The value of the information criterion as an objective function with respect to the tuning parameter \code{lamdba}.
}
\description{
The objective function (an information criterion) for optimization in WOCR models.
}
\details{
This function computes the value of the objective function for the tuning parameter \code{lamdba},
depending on which WOCR model is used. See Su et al. (2016+) for the recommended choices. Different models have different weighting strategies and
different degrees of freedom. This function is not intended for use outside of this package.
}
\references{
\itemize{
\item Su, X., Wonkye, Y., Wang, P., and Yin, X. (2016+). Weighted orthogonal component regression (WOCR).
Submitted.
}
}
\seealso{
\code{\link[WOCR]{WOCR}}, \code{\link[WOCR]{weighting}}
}
