% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-WOCR.R
\name{plot.WOCR}
\alias{plot.WOCR}
\alias{print.wocr}
\title{The Generic \code{plot} Function for Object of \code{WOCR} Class}
\usage{
\method{plot}{WOCR}(fit, horizontal = TRUE, mar = rep(4, 4), ...)
}
\arguments{
\item{fit}{an object of \code{WOCR} class.}

\item{horizontal}{Logical indicator of horizontal alignment. Default is \code{TRUE}.}

\item{mar}{margin in terms of the number of lines to be specified on the four sides of the plot}

\item{\code{...}}{further arguments passed to or from other methods.}
}
\value{
Plots of the singular values \eqn{d}, the coefficients \eqn{gamma} for components, and the
estimated weigt for each components, versus the components.
}
\description{
The Generic \code{plot} Function for Object of \code{WOCR} Class
}
\details{
The (generic) plot method for an \code{WOCR} object. It generates three plots of \eqn{d}, \eqn{gamma}, and weight.
}
\references{
\itemize{
\item Su, X., Wonkye, Y., Wang, P., and Yin, X. (2016+). Weighted orthogonal component regression (WOCR).
Submitted.
}
}
\seealso{
\code{\link[WOCR]{WOCR}}
}
