% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-WOCR.R
\name{predict.WOCR}
\alias{predict.WOCR}
\alias{predict.wocr}
\title{The Generic \code{predict} Function for Object of \code{WOCR} Class}
\usage{
\method{predict}{WOCR}(fit, newdata = NULL)
}
\arguments{
\item{fit}{an object of \code{WOCR} class.}

\item{newdata}{An optional data frame in which to look for variables with which to predict. If omitted, the fitted values
(based on the original WOCR fit) are returned.}
}
\value{
The predicted values for observations in the new data.
}
\description{
The Generic \code{predict} Function for Object of \code{WOCR} Class
}
\examples{
   data(BostonHousing1)
   fit.WOCR <- WOCR(formula=cmedv~., data=BostonHousing1, model="RR.d.lambda")
   # THE SAME DATA IS USED FOR PREDICTION HERE JUST FOR ILLUSTRATION
   par(mfrow=c(1, 1), mar=rep(4, 4))
   y.pred <- predict(fit.WOCR, newdata=BostonHousing1)
   plot(BostonHousing1$cmedv, y.pred, pch=19, cex=.6, col="cadetblue3", xlab="observed", ylab="predicted")
   abline(a=0, b=1, col="chocolate1", lwd=1.8)

}
\references{
\itemize{
\item Su, X., Wonkye, Y., Wang, P., and Yin, X. (2016+). Weighted orthogonal component regression (WOCR).
Submitted.
}
}
\seealso{
\code{\link[WOCR]{WOCR}}
}
