% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-WOCR.R
\name{print.WOCR}
\alias{print.WOCR}
\alias{print.wocr}
\title{The Generic \code{print} Function for Object of \code{WOCR} Class}
\usage{
\method{print}{WOCR}(fit, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{fit}{an object of \code{WOCR} class.}

\item{digits}{the minimal number of significant digits. See \code{\link[base]{print.default}}.}

\item{\code{...}}{further arguments passed to or from other methods.}
}
\value{
A table with four columns (\code{d},  \code{gamma}, \code{w.hat}, \code{beta.hat}).
}
\description{
The Generic \code{print} Function for Object of \code{WOCR} Class
}
\details{
The (generic) print method for an \code{WOCR} object. The results include \code{d} (the singular value of standardized
\eqn{X}), \code{gamma} (the coefficients for the orthogonal components), \code{w.hat} (the estimated weights),
\code{beta.hat} (the estimated coefficients for the standardized \eqn{X} variables).
}
\references{
\itemize{
\item Su, X., Wonkye, Y., Wang, P., and Yin, X. (2016+). Weighted orthogonal component regression (WOCR).
Submitted.
}
}
\seealso{
\code{\link[WOCR]{WOCR}}
}
