% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WOCR.R
\name{WOCR}
\alias{WOCR}
\alias{wocr}
\title{Fitting the weighted orthoginal components regression (WOCR) models}
\usage{
WOCR(formula, data = NULL, scale = TRUE, model = "RR.d.lambda", a = 20,
  criterion = NULL, use.GenSA = TRUE, maxit.global = 100, eps = 1e-06,
  LB = NULL, UB = NULL, details = FALSE)
}
\arguments{
\item{formula}{An object of class \code{\link[stats]{formula}}, with the response on the left of
a \code{~} operator, and the terms on the right. The function assumes the intercept to be zero with scaled data.
If the interacept \code{~ +1} is included explicitly or by default, it will  be automatically removed.}

\item{data}{A data.frame in which to interpret the variables named in the \code{formula} argument.}

\item{scale}{Logical indicator of whether you want to scale the data. Here, \code{scale=TRUE} should always be done unless the
data have already scaled or standardized. This action involves standaradizing the columns in matrix \eqn{X}  and center the
response vector \eqn{y}. Note that this option also has an effect on the resultant prediction.}

\item{model}{Specifies the WOCR model to be fitted. Six choices are possible: \code{RR.d.lambda},
\code{RR.gamma.lambda}, \code{PCR.d.c}, \code{PCR.gamma.c}, \code{PCR.d.a.c} and \code{PCR.gamma.a.c}.}

\item{a}{The fixed shape parameter \eqn{a} in models \code{PCR.d.c} and \code{PCR.gamma.c} only. By default, \eqn{a} is set to 20.}

\item{criterion}{Specifies the information criterion to be minimized. The appropriate choice is one from \code{"AIC"}, \code{"BIC"}, or \code{"GCV"}.
By default, \code{criterion} is left as \code{NULL} and the recommended one is used.}

\item{use.GenSA}{Logical value indicating if the generalized simulated annealing \code{GenSA} is used for optimizing the
2-D optimization problems in models \code{PCR.d.a.c} and \code{PCR.gamma.a.c}. The default is \code{TRUE}.}

\item{maxit.global}{Maximum number of iterations allowed for the global optimization algorithm, which is either \code{GenSA} or
\code{SANN} in \code{\link[stats]{optim}}. Default value is 100.}

\item{LB}{The lower bounds for the search space in \code{GenSA} The default is \code{NULL} and recommended one is used.}

\item{UB}{The upper bounds for the search space in \code{GenSA} The default is \code{NULL} and recommended one is used.}

\item{details}{Logical value: if \code{TRUE}, detailed results will be printed out when running \code{coxphMIC}.}

\item{epsilon}{Tolerance level for convergence. Default is 1e-6.}

\item{maxit.local}{Maximum number of iterations allowed for the local optimizaiton algorithm \code{BFGS}. Default value is 100.}
}
\value{
An object of class \code{WOCR} is returned, which may contain the following components depending on the options.
\describe{
\item{call}{the matched call.}
\item{opt.global}{Results from the preliminary run of a global optimization procedure (\code{SANN} as default or \code{GenSA}).}
\item{opt}{Results from the final optimization procedure. This could be the results from the 1-D optimization \code{\link[stats]{optimize}} or from the second run of the local optimizaiton \code{BFGS}.}
\item{m}{The rank of the design matrix \eqn{X}.}
\item{d}{The vector of singular values.}
\item{gamma}{The vector of coefficients, consisting of the inner product of each orthogonal component and the repsonse \eqn{y}.}
\item{formula}{The \code{formula} used in fitting WOCR. A copy of \code{formula} will be useful for prediction later on.}
\item{mean.y}{Mean value of the response \eqn{y}.}
\item{mean.x}{Mean vector of columns of matrix \eqn{X}.}
\item{sd.x}{Vector of standard deviations for each \eqn{X} variables or columns.}
\item{min.Q}{Value of the minimized objective function.}
\item{lambda.hat}{The 1-D or 2-D estimated tuning parameter(s).}
\item{beta.hat}{The coefficients for the \emph{standardized} \eqn{X} variables, instead of the orthogonal components.}
\item{w.hat}{The estimated weights.}
\item{n.comp}{The selected number of components for WOCR models \code{PCR.d.c} and \code{PCR.gamma.c} only.}
\item{y.hat}{The fitted value for the response. The mean response is added back. So \code{y.hat} can be compared directly to the response \eqn{y}.}
}
}
\description{
Fitting the weighted orthoginal components regression (WOCR) models
}
\details{
The main idea of WOCR is to parameterize the weights for orthogonal components with a simple function whose specification is up to one or two
tuning parameters. The weighting strategy takes advantage of the inherent monotonicity among these orthogonal components.

To solve the 2-D smooth yet nonconvex optimization, two options are made available. The first is a simulated annealing (\code{method="SANN"} option
in \code{\link[stats]{optim}}) global optimization algorithm is first applied. The resultant estimator is then used
as the starting point for another local optimization algorithm, where the quasi-Newton BFGS method (\code{method="BFGS"}
in \code{\link{optim}}) by default. Optionally, the generalized simulated annealing, implemented in \code{\link[GenSA]{GenSA}},
can be used instead. This latter approach tends to be slower. However, it does not need to be combined with another local optimization;
besides, it often yields the same final solution with different runs. Thus, when \code{use.GenSA=TRUE},
the output includes \code{opt.global} only, without \code{opt.local}.
}
\examples{
   data(BostonHousing1)
   fit.wocr <- WOCR(formula=cmedv~., data=BostonHousing1, model="RR.d.lambda")
   print(fit.wocr)
   plot(fit.wocr, horizontal=TRUE)

}
\references{
\itemize{
\item Su, X., Wonkye, Y., Wang, P., and Yin, X. (2016+). Weighted orthogonal component regression (WOCR).
Submitted.
}
}
\seealso{
\code{\link[pls]{pcr}}
}
